3_Finetuning/requirements.txt
# Core deep learning
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.5.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Logging and monitoring
tensorboard>=2.13.0
tqdm>=4.65.0

# Model export
onnx>=1.14.0
onnxruntime>=1.15.0

# Jupyter
jupyter>=1.0.0
ipywidgets>=8.0.0

# Dataset download
kaggle>=1.5.0
requests>=2.31.0

# Utils
pyyaml>=6.0
